Refactor and fix broken import
