+++ src/barrier/__tests__/extractBarrier-test.js
@@ -1,0 +1,1 @@
+
